pulling project details content into it's own component
