+++ src/components/ProjectDetails/ProjectDetails.js
@@ -5,6 +5,7 @@ import { transformProjectDetails } from "../../utils";
 import { Card, CardContent, LinearProgress, Typography } from "@mui/material";
 import ProjectDetailsDivider from "../ProjectDetailsDivider";
 import ProjectDetailsHeader from "../ProjectDetailsHeader";
+import ProjectDetailsContent from "../ProjectDetailsContent";
 
 export default function ProjectDetails() {
   const { address } = useParams();
@@ -37,60 +38,31 @@ export default function ProjectDetails() {
         subheader={`Started By: ${fundStarter}`}
         subheaderTypographyProps={{ sx: { wordBreak: "break-all" } }}
       />
-      <CardContent>
-        {projectAddress && (
-          <>
-            <ProjectDetailsDivider>Project Address</ProjectDetailsDivider>
-            <Typography sx={{ wordBreak: "break-all" }}>
-              {projectAddress}
-            </Typography>
-          </>
-        )}
-        {goal && currentAmount && (
-          <>
-            <ProjectDetailsDivider>Campaign</ProjectDetailsDivider>
-            <Typography>
-              Raised {currentAmount} of {goal} ETH
-            </Typography>
-            <LinearProgress
-              variant="determinate"
-              value={
-                Number(currentAmount)
-                  ? (goal / currentAmount) * 100
-                  : Number(currentAmount)
-              }
-            />
-          </>
-        )}
-        {fundDescription && (
-          <>
-            <ProjectDetailsDivider>Description</ProjectDetailsDivider>
-            <Typography>{fundDescription}</Typography>
-          </>
-        )}
-        {deadline && (
-          <Typography
-            sx={{
-              textAlign: "right",
-              fontStyle: "italic",
-              color: (theme) => theme.palette.info.main,
-            }}
-          >
-            Deadline: {deadline}
-          </Typography>
-        )}
-        {(currentState || currentState === 0) && (
-          <Typography
-            sx={{
-              textAlign: "right",
-              fontStyle: "italic",
-              color: (theme) => theme.palette.info.main,
-            }}
-          >
-            Status: {currentState}
-          </Typography>
-        )}
-      </CardContent>
+      <ProjectDetailsContent
+        {...{ projectAddress, goal, currentAmount, fundDescription }}
+      />
+      {deadline && (
+        <Typography
+          sx={{
+            textAlign: "right",
+            fontStyle: "italic",
+            color: (theme) => theme.palette.info.main,
+          }}
+        >
+          Deadline: {deadline}
+        </Typography>
+      )}
+      {(currentState || currentState === 0) && (
+        <Typography
+          sx={{
+            textAlign: "right",
+            fontStyle: "italic",
+            color: (theme) => theme.palette.info.main,
+          }}
+        >
+          Status: {currentState}
+        </Typography>
+      )}
     </Card>
   );
 }
